+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/LibsDisguiseManager.java
@@ -1,13 +1,16 @@
 package com.elmakers.mine.bukkit.integration;
 
 import java.util.logging.Level;
+import javax.annotation.Nullable;
 
 import org.bukkit.Material;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.Player;
 import org.bukkit.entity.Villager;
 import org.bukkit.plugin.Plugin;
 
+import com.comphenix.protocol.wrappers.WrappedGameProfile;
 import com.elmakers.mine.bukkit.utility.DeprecatedUtils;
 
 import me.libraryaddict.disguise.DisguiseAPI;
@@ -20,6 +23,7 @@ import me.libraryaddict.disguise.disguisetypes.MobDisguise;
 import me.libraryaddict.disguise.disguisetypes.PlayerDisguise;
 import me.libraryaddict.disguise.disguisetypes.watchers.ArmorStandWatcher;
 import me.libraryaddict.disguise.disguisetypes.watchers.VillagerWatcher;
+import me.libraryaddict.disguise.utilities.DisguiseUtilities;
 
 public class LibsDisguiseManager {
     private final Plugin disguisePlugin;
@@ -81,7 +85,6 @@ public class LibsDisguiseManager {
                     disguise = paintingDisguise;
                     break;
                 case ARROW:
-                case TIPPED_ARROW:
                 case SPECTRAL_ARROW:
                 case FIREBALL:
                 case SMALL_FIREBALL:
@@ -123,5 +126,14 @@ public class LibsDisguiseManager {
             return false;
         }
         return true;
+    }
+
+    @Nullable
+    public String getSkin(Player player) {
+        WrappedGameProfile profile = WrappedGameProfile.fromPlayer(player);
+        if (profile == null) {
+            return null;
+        }
+        return DisguiseUtilities.getGson().toJson(profile);
     }
 }