+++ Magic/CHANGELOG.md
@@ -4,7 +4,8 @@
 
  - Add BattleArena integration, friendly fire flags will respect teams
  - Add magic_sp, magic_spell_count, magic_mana and magic_mana_max Placeholders
- - More graceful hand
+ - More graceful handling of invalid wand icons
+ - Add "/mage skin" command, which dumps a block of text you can put in a "skin" property of a player disguise
 
 # 7.8.4